minor fix
Just saw control treatment for Loxo 2 was missing. Found the cause and also realized the first row of all trait measurements was missing due to me using na.omit in combination with a fulljoin. --> pcgr is always NA for the first day when lag.x was used. Fixed this and updated the csv files.